requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.3
email-validator==2.1.0.post1
requests==2.31.0
simple-salesforce==1.12.4
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0